move logic to file. Add comments. Add prepared statements
